routes and prisma functions updated

+++ src/users/users.service.ts
@@ -1,20 +1,17 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { User, Prisma } from '@prisma/client';
 import { PrismaService } from '../prisma.service';
+import { CreateUserDto } from './dto/create-user.dto';
+import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UsersService {
   constructor(private prisma: PrismaService) {}
 
-  async create(data: Prisma.UserCreateInput): Promise<User> {
-    try {
-      const user = this.prisma.user.create({
-        data,
-      });
-      return user;
-    } catch (err) {
-      throw new HttpException('Create Error', HttpStatus.INTERNAL_SERVER_ERROR);
-    }
+  async create(data: CreateUserDto): Promise<User> {
+    return this.prisma.user.create({
+      data,
+    });
   }
 
   async findAll(params: {
@@ -48,17 +45,29 @@ export class UsersService {
 
   async update(params: {
     where: Prisma.UserWhereUniqueInput;
-    data: Prisma.UserUpdateInput;
+    data: UpdateUserDto;
   }): Promise<User> {
     const { where, data } = params;
+    const existingUser = await this.prisma.user.findUnique({
+      where,
+    });
+    if (!existingUser) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
     return this.prisma.user.update({
       data,
       where,
     });
   }
 
-  async remove(where: Prisma.UserWhereUniqueInput): Promise<User> {
-    return this.prisma.user.delete({
+  async remove(where: Prisma.UserWhereUniqueInput) {
+    const existingUser = await this.prisma.user.findUnique({
+      where,
+    });
+    if (!existingUser) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    await this.prisma.user.delete({
       where,
     });
   }
